requirements.txt
# RAG系统依赖包
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-openai>=0.0.5
langchain-huggingface>=0.0.5
faiss-cpu>=1.7.4
huggingface-hub>=0.19.0
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
PyYAML>=6.0
markdown>=3.5
unstructured>=0.10.30
python-docx>=0.8.11
openai>=1.3.0
streamlit>=1.25.0 